protect deleteVideo

+++ src/controllers/videoController.js
@@ -83,11 +83,15 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
-    const { id } = req.params
+    const { id } = req.params;
+    const {user: {_id}} = req.session;
     const video = await Video.findById(id)
     if(!video){
         return res.status(404).render("404", {pageTitle: "Video is not found"})
     };
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/")
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect('/')
 }
